move StatusBarOnFocus to src/components
