+++ src/components/status-bar-on-focus/index.tsx
