+++ src/screens/Map/MapScreen.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Map} from '@atb/components/map';
 import useIsScreenReaderEnabled from '@atb/utils/use-is-screen-reader-enabled';
 import {MapDisabledForScreenReader} from './components/MapDisabledForScreenReader';
-import StatusBarOnFocus from './components/StatusBarOnFocus';
+import StatusBarOnFocus from '@atb/components/status-bar-on-focus';
 
 export function MapScreen() {
   const isScreenReaderEnabled = useIsScreenReaderEnabled();
